mcc_driver

Trait Callbacks

pub trait Callbacks {
    type Output;

    // Provided methods
    fn after_parse<'db>(
        &mut self,
        _db: &'db dyn Db,
        _source_file: SourceFile,
        _ast: Ast<'db>,
        _diags: Vec<&Diagnostics>,
    ) -> ControlFlow<Self::Output> { ... }
    fn after_lower<'db>(
        &mut self,
        _db: &'db dyn Db,
        _tacky: Program<'db>,
        _diags: Vec<&Diagnostics>,
    ) -> ControlFlow<Self::Output> { ... }
    fn after_codegen<'db>(
        &mut self,
        _db: &'db dyn Db,
        _asm: Program<'db>,
        _diags: Vec<&Diagnostics>,
    ) -> ControlFlow<Self::Output> { ... }
    fn after_render_assembly(
        &mut self,
        _db: &dyn Db,
        _asm: Text,
        _diags: Vec<&Diagnostics>,
    ) -> ControlFlow<Self::Output> { ... }
    fn after_compile(
        &mut self,
        _db: &dyn Db,
        _binary: PathBuf,
    ) -> ControlFlow<Self::Output> { ... }
}

Callbacks fired at various stages of compilation.

The callbacks are called in the following order:

after_parse

after_lower

after_codegen

after_render_assembly

after_compile

If a callback returns ControlFlow::Break, the compiler will stop running and return early.

Required Associated Types

type Output

Provided Methods

fn after_parse<'db>( &mut self, _db: &'db dyn Db, _source_file: SourceFile, _ast: Ast<'db>, _diags: Vec<&Diagnostics>, ) -> ControlFlow<Self::Output>

Called after parsing the file.

fn after_lower<'db>( &mut self, _db: &'db dyn Db, _tacky: Program<'db>, _diags: Vec<&Diagnostics>, ) -> ControlFlow<Self::Output>

Called after lowering the Abstract Syntax Tree to Three Address Code.

fn after_codegen<'db>( &mut self, _db: &'db dyn Db, _asm: Program<'db>, _diags: Vec<&Diagnostics>, ) -> ControlFlow<Self::Output>

Called after generating assembly instructions from Three Address Code.

fn after_render_assembly( &mut self, _db: &dyn Db, _asm: Text, _diags: Vec<&Diagnostics>, ) -> ControlFlow<Self::Output>

Called after rendering the assembly instructions to a string.

fn after_compile( &mut self, _db: &dyn Db, _binary: PathBuf, ) -> ControlFlow<Self::Output>